requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.5.0
pycountry>=22.3.5
python-dotenv>=0.19.0
numpy>=1.23.0